requirements.txt
asgiref==3.9.2
astroid==3.3.11
asttokens==3.0.0
colorama==0.4.6
decorator==5.2.1
dill==0.4.0
Django==5.2.6
django-extensions==4.1
django-js-asset==3.1.2
django-mptt==0.18.0
executing==2.2.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
matplotlib-inline==0.1.7
mccabe==0.7.0
packaging==25.0
parso==0.8.5
platformdirs==4.4.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
pylint==3.3.8
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
sqlparse==0.5.3
stack-data==0.6.3
tomlkit==0.13.3
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.14
